requirements.txt
# Core LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0

# Ollama integration
langchain-ollama>=0.2.0

# Vector database
chromadb>=0.5.23

# PDF processing
pymupdf>=1.25.0
pypdf>=4.0.0

# API and Frontend
fastapi>=0.115.0
uvicorn>=0.34.0
streamlit>=1.41.0
httpx>=0.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0

# Notebook support
ipykernel>=6.29.0
ipywidgets>=8.1.0
